UCT: CR fixes

+++ src/uct/ib/base/ib_device.h
@@ -411,7 +411,7 @@ ucs_status_t uct_ib_device_get_roce_ndev_name(uct_ib_device_t *dev,
 
 ucs_status_t
 uct_ib_device_get_roce_ndev_index(uct_ib_device_t *dev, uint8_t port_num,
-                                  int gid_index, int *ndev_index_p);
+                                  uint8_t gid_index, int *ndev_index_p);
 
 unsigned uct_ib_device_get_roce_lag_level(uct_ib_device_t *dev,
                                           uint8_t port_num,
